requirements.txt
pytelegrambotapi
sqlalchemy
alembic
psycopg2-binary
python-dotenv
beautifulsoup4
lxml
python-telegram-bot-calendar
reportlab
qrcode[pil]
PyPDF2
segno
httpx
fuzzywuzzy
python-levenshtein
loguru
celery[redis]
celery-beat